+++ src/renderer/src/components/task-page-github-issue-creation.test.ts
@@ -5,6 +5,10 @@ const taskPageSource = readFileSync(
   new URL('./task-page/hooks/use-task-page-create-github-submit.ts', import.meta.url),
   'utf8'
 )
+const newIssueStateSource = readFileSync(
+  new URL('./task-page/hooks/use-task-page-github-new-issue-state.ts', import.meta.url),
+  'utf8'
+)
 
 function issueCreationSection(): string {
   const start = taskPageSource.indexOf('const handleCreateNewIssue')
@@ -15,6 +19,11 @@ function issueCreationSection(): string {
 }
 
 describe('TaskPage GitHub issue creation', () => {
+  it('keeps issue creation targeted to the first selected repo on a fresh mount', () => {
+    expect(newIssueStateSource).toContain('(selectedRepos[0]?.id ?? null)')
+    expect(newIssueStateSource).toContain('newIssueRepoId !== null')
+  })
+
   it('covers the complete remote oversized-body recovery timeout envelope', () => {
     const section = issueCreationSection()
