+++ src/renderer/src/components/task-page/dialogs/new-linear-issue-more-pickers.tsx
@@ -57,7 +57,7 @@ export function NewLinearIssueMorePickers({
             <ChevronDown className="size-3 text-muted-foreground/70" />
           </button>
         </PopoverTrigger>
-        <PopoverContent align="start" className="w-48 p-1">
+        <PopoverContent align="start" className="w-48 p-1 popover-scroll-content scrollbar-sleek">
           <div className="text-[10px] font-semibold text-muted-foreground px-2 py-1 uppercase tracking-wider">
             {translate('auto.components.TaskPage.c8d5bec5f7', 'Priority')}
           </div>
@@ -109,7 +109,7 @@ export function NewLinearIssueMorePickers({
             <ChevronDown className="size-3 text-muted-foreground/70" />
           </button>
         </PopoverTrigger>
-        <PopoverContent align="start" className="w-64 p-1">
+        <PopoverContent align="start" className="w-64 p-1 popover-scroll-content scrollbar-sleek">
           <div className="text-[10px] font-semibold text-muted-foreground px-2 py-1 uppercase tracking-wider">
             {translate('auto.components.TaskPage.00022ec0ba', 'Project')}
           </div>
@@ -118,7 +118,7 @@ export function NewLinearIssueMorePickers({
               <LoaderCircle className="size-4 animate-spin text-muted-foreground" />
             </div>
           ) : (
-            <div className="max-h-60 overflow-y-auto scrollbar-sleek">
+            <div>
               <button
                 type="button"
                 onClick={() => setNewLinearIssueProjectId(null)}
@@ -176,7 +176,7 @@ export function NewLinearIssueMorePickers({
             <ChevronDown className="size-3 text-muted-foreground/70" />
           </button>
         </PopoverTrigger>
-        <PopoverContent align="start" className="w-64 p-1">
+        <PopoverContent align="start" className="w-64 p-1 popover-scroll-content scrollbar-sleek">
           <div className="text-[10px] font-semibold text-muted-foreground px-2 py-1 uppercase tracking-wider">
             {translate('auto.components.TaskPage.d0ca4aa1d0', 'Labels')}
           </div>
@@ -185,7 +185,7 @@ export function NewLinearIssueMorePickers({
               <LoaderCircle className="size-4 animate-spin text-muted-foreground" />
             </div>
           ) : (
-            <div className="max-h-60 overflow-y-auto scrollbar-sleek">
+            <div>
               {newLinearLabels.data.map((l) => {
                 const isSelected = newLinearIssueLabelIds.includes(l.id)
                 return (
