+++ src/renderer/src/components/task-page/dialogs/new-linear-issue-status-assignee.tsx
@@ -55,7 +55,7 @@ export function NewLinearIssueStatusAssignee({
             <ChevronDown className="size-3 text-muted-foreground/70" />
           </button>
         </PopoverTrigger>
-        <PopoverContent align="start" className="w-56 p-1">
+        <PopoverContent align="start" className="w-56 p-1 popover-scroll-content scrollbar-sleek">
           <div className="text-[10px] font-semibold text-muted-foreground px-2 py-1 uppercase tracking-wider">
             {translate('auto.components.TaskPage.154b0fa623', 'Status')}
           </div>
@@ -64,7 +64,7 @@ export function NewLinearIssueStatusAssignee({
               <LoaderCircle className="size-4 animate-spin text-muted-foreground" />
             </div>
           ) : (
-            <div className="max-h-60 overflow-y-auto scrollbar-sleek">
+            <div>
               {newLinearStates.data.map((s) => (
                 <button
                   key={s.id}
@@ -128,7 +128,7 @@ export function NewLinearIssueStatusAssignee({
             <ChevronDown className="size-3 text-muted-foreground/70" />
           </button>
         </PopoverTrigger>
-        <PopoverContent align="start" className="w-64 p-1">
+        <PopoverContent align="start" className="w-64 p-1 popover-scroll-content scrollbar-sleek">
           <div className="text-[10px] font-semibold text-muted-foreground px-2 py-1 uppercase tracking-wider">
             {translate('auto.components.TaskPage.d2a876ca53', 'Assignee')}
           </div>
@@ -137,7 +137,7 @@ export function NewLinearIssueStatusAssignee({
               <LoaderCircle className="size-4 animate-spin text-muted-foreground" />
             </div>
           ) : (
-            <div className="max-h-60 overflow-y-auto scrollbar-sleek">
+            <div>
               <button
                 type="button"
                 onClick={() => setNewLinearIssueAssigneeId(null)}
